1.0.3-Alpha7
If a user starts to install the app, but backs out before completing configuration apparently the SmartThings environment installs the app, unconfigured.  This leads to multiple instances and none of these instances have an Uninstall button on the Configure page as expected.  This release ensures the Uninstall button is always on the Configure page to address this issue.